Effective Navigation Of Ambulance
This Project rectifies the most Important Problem in India, That is Ambulance getting stuck in traffic. According to the research done by our team, more than 30% of our people are dying due to ambulance stuck in traffic. So our Aim is to reduce the death rate using this Effective Navigation of Ambulances.